Fix: Thrown out old or stale code
Most of these will come back at some point, but for version `v2.0.0` these aren't needed
