--- src/AthenaLib/HTML/data/css_selection_type.py
@@ -1,20 +1,0 @@
-# ----------------------------------------------------------------------------------------------------------------------
-# - Package Imports -
-# ----------------------------------------------------------------------------------------------------------------------
-# General Packages
-from __future__ import annotations
-import enum
-
-# Custom Library
-
-# Custom Packages
-
-# ----------------------------------------------------------------------------------------------------------------------
-# - Code -
-# ----------------------------------------------------------------------------------------------------------------------
-class CSSSelectionType(enum.Enum):
-    default=""
-    inside=" "
-    family=">"
-    following="+"
-    preceding="~"
