--- src/AthenaLib/data/types.py
@@ -1,17 +1,0 @@
-# ----------------------------------------------------------------------------------------------------------------------
-# - Package Imports -
-# ----------------------------------------------------------------------------------------------------------------------
-# General Packages
-from __future__ import annotations
-import os
-
-# Custom Library
-
-# Custom Packages
-
-# ----------------------------------------------------------------------------------------------------------------------
-# - Code -
-# ----------------------------------------------------------------------------------------------------------------------
-PATHLIKE = str|bytes|os.PathLike
-NUMBER = float|int
-POINT = tuple[NUMBER, NUMBER] | list[NUMBER, NUMBER]
