--- src/AthenaLib/functions/string.py
@@ -1,18 +1,0 @@
-# ----------------------------------------------------------------------------------------------------------------------
-# - Package Imports -
-# ----------------------------------------------------------------------------------------------------------------------
-# General Packages
-from __future__ import annotations
-
-# Custom Library
-
-# Custom Packages
-
-# ----------------------------------------------------------------------------------------------------------------------
-# - Code -
-# ----------------------------------------------------------------------------------------------------------------------
-def enclose_double_quote(text:str) -> str:
-    return f'"{text}"'
-
-def enclose_single_quote(text:str) -> str:
-    return f"'{text}'"
